backend/requirements.txt
# Web Framework & Utils
Flask==3.1.2
flask-cors==6.0.1
python-dotenv==1.0.1
werkzeug==3.1.6
gunicorn==21.2.0

# Firebase & Cloud
firebase_admin==7.1.0

# AI & Processing
groq==0.13.0
rapidfuzz==3.6.1

# File Parsers
pypdf==6.8.0
python-docx==1.2.0

# Testing & Utilities
requests==2.32.5

